+++ Pandora/Paradigm/Primary/Algebraic.hs
@@ -4,6 +4,7 @@ module Pandora.Paradigm.Primary.Algebraic (module Exports, (-<*>-)) where
 
 import Pandora.Paradigm.Primary.Algebraic.Exponential as Exports
 import Pandora.Paradigm.Primary.Algebraic.Product as Exports
+import Pandora.Paradigm.Primary.Algebraic.Sum as Exports
 
 import Pandora.Pattern.Category (($))
 import Pandora.Pattern.Functor.Covariant (Covariant_ ((-<$>-)))