+++ Pandora/Paradigm/Primary/Algebraic/Sum.hs
@@ -1,0 +1,5 @@
+module Pandora.Paradigm.Primary.Algebraic.Sum where
+
+infixr 0 :+:
+
+data (:+:) s a = Option s | Adoption a